+++ src/pages/notification/NotificationHeader.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import previous_icon from "../../../assets/icons/arrow-left.svg";
+import previous_icon from "../../assets/icons/arrow-left.svg";
 import { useNavigate } from 'react-router-dom';
